Use raw query string to retrieve patient guid in place of parameter attribute

+++ UI/Areas/PatientCaseReview/Pages/ViewModels/ClinicalDataViewModel.cs
@@ -3,8 +3,8 @@
     public class ClinicalDataViewModel
     {
         public Guid PatientGuid { get; set; }
-        public int Temperature { get; set; }
-        public int HgCount { get; set;} 
+        public double Temperature { get; set; }
+        public double HgCount { get; set;} 
         public DateTime? LastVisitDateTime { get; set; } 
         public string PrescriptionValues { get; set; } = string.Empty; // Use single string for now
         //public string[]? LastPrescriptions { get; set;} = PrescriptionValues.Split(',');
